+++ pkg/lightning/tikv/local_sst_writer.go
@@ -15,126 +15,17 @@
 package tikv
 
 import (
-	"context"
 	"encoding/binary"
 
-	"github.com/google/uuid"
 	rocks "github.com/lance6716/pebble"
 	rocksbloom "github.com/lance6716/pebble/bloom"
 	"github.com/lance6716/pebble/objstorage/objstorageprovider"
 	rockssst "github.com/lance6716/pebble/sstable"
 	"github.com/lance6716/pebble/vfs"
 	"github.com/pingcap/errors"
-	"github.com/pingcap/kvproto/pkg/import_sstpb"
-	"github.com/pingcap/kvproto/pkg/kvrpcpb"
 	"github.com/pingcap/tidb/pkg/util/codec"
 	"github.com/pingcap/tidb/pkg/util/intest"
-	pd "github.com/tikv/pd/client"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
 )
-
-// write2ImportService4Test writes these sorted key-value pairs to the TiKV
-// cluster. SST files are generated by TiKV and saved in import directory if no
-// error happens.
-func write2ImportService4Test(
-	ctx context.Context,
-	pdAddrs []string,
-	sortedKVs [][2][]byte,
-	ts uint64,
-) ([]*import_sstpb.SSTMeta, error) {
-	pdClient, err := pd.NewClient(pdAddrs, pd.SecurityOption{})
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	defer pdClient.Close()
-
-	r0, err := pdClient.GetRegion(ctx, sortedKVs[0][0])
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	r1, err := pdClient.GetRegion(ctx, sortedKVs[len(sortedKVs)-1][0])
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	if r0.Meta.Id != r1.Meta.Id {
-		return nil, errors.Errorf(
-			"only support write to the same region, "+
-				"first key: %X, last key: %X, "+
-				"first region id: %d, last region id: %d",
-			sortedKVs[0][0], sortedKVs[len(sortedKVs)-1][0],
-			r0.Meta.Id, r1.Meta.Id,
-		)
-	}
-
-	store, err := pdClient.GetStore(ctx, r0.Leader.GetStoreId())
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	conn, err := grpc.DialContext(
-		ctx, store.GetAddress(),
-		grpc.WithTransportCredentials(insecure.NewCredentials()),
-		grpc.WithBlock(),
-	)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	defer conn.Close()
-
-	ingestClient := import_sstpb.NewImportSSTClient(conn)
-	writeStream, err := ingestClient.Write(ctx)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	u := uuid.New()
-	writeMeta := &import_sstpb.SSTMeta{
-		Uuid:        u[:],
-		RegionId:    r0.Meta.Id,
-		RegionEpoch: r0.Meta.RegionEpoch,
-		Range: &import_sstpb.Range{
-			Start: sortedKVs[0][0],
-			End:   sortedKVs[len(sortedKVs)-1][0],
-		},
-	}
-	rpcCtx := kvrpcpb.Context{
-		RegionId:    r0.Meta.Id,
-		RegionEpoch: r0.Meta.RegionEpoch,
-		Peer:        r0.Leader,
-	}
-	err = writeStream.Send(&import_sstpb.WriteRequest{
-		Chunk:   &import_sstpb.WriteRequest_Meta{Meta: writeMeta},
-		Context: &rpcCtx,
-	})
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	batch := &import_sstpb.WriteBatch{
-		CommitTs: ts,
-		Pairs:    make([]*import_sstpb.Pair, 0, len(sortedKVs)),
-	}
-	for _, kv := range sortedKVs {
-		batch.Pairs = append(batch.Pairs, &import_sstpb.Pair{
-			Key:   kv[0],
-			Value: kv[1],
-		})
-	}
-	err = writeStream.Send(&import_sstpb.WriteRequest{
-		Chunk:   &import_sstpb.WriteRequest_Batch{Batch: batch},
-		Context: &rpcCtx,
-	})
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	resp, err := writeStream.CloseAndRecv()
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	if resp.GetError() != nil {
-		return nil, errors.Errorf("write failed: %s", resp.GetError())
-	}
-	return resp.Metas, nil
-}
 
 // writeCFWriter generates SST files for TiKV's write column family.
 type writeCFWriter struct {
@@ -176,17 +67,32 @@ func newWriteCFWriter(
 	return &writeCFWriter{sstWriter: writer, ts: ts}, nil
 }
 
-// set mimic TiKV's `TxnSstWriter.put`. ref
-// https://github.com/tikv/tikv/blob/7793f1d5dc40206fe406ca001be1e0d7f1b83a8f/components/sst_importer/src/sst_writer.rs#L92.
+// set mimic TiKV's TxnSstWriter logic to encode key and value and write to SST.
 func (w *writeCFWriter) set(key, value []byte) error {
 	intest.Assert(isShortValue(value), "not implemented, need to write to default CF")
 
-	actualKey := make([]byte, 0, 1+len(key)+8)
+	actualKey := make([]byte, 0, 1+codec.EncodedBytesLength(len(key))+8)
+	// keys::data_key will add the 'z' prefix [1] at `TxnSstWriter.put` [2].
+	//
+	// [1] https://github.com/tikv/tikv/blob/7793f1d5dc40206fe406ca001be1e0d7f1b83a8f/components/keys/src/lib.rs#L206
+	// [2] https://github.com/tikv/tikv/blob/7793f1d5dc40206fe406ca001be1e0d7f1b83a8f/components/sst_importer/src/sst_writer.rs#L92
 	actualKey = append(actualKey, 'z')
+	// Key::from_raw [3] will encode the key as bytes at `TxnSstWriter.write` [4],
+	// which is the caller of `TxnSstWriter.put` [2].
+	//
+	// [3] https://github.com/tikv/tikv/blob/7793f1d5dc40206fe406ca001be1e0d7f1b83a8f/components/txn_types/src/types.rs#L55
+	// [4] https://github.com/tikv/tikv/blob/7793f1d5dc40206fe406ca001be1e0d7f1b83a8f/components/sst_importer/src/sst_writer.rs#L74
 	actualKey = codec.EncodeBytes(actualKey, key)
+	// Key::append_ts [5] will append the bit-wise reverted ts at
+	// `TxnSstWriter.write` [4].
+	//
+	// [5] https://github.com/tikv/tikv/blob/7793f1d5dc40206fe406ca001be1e0d7f1b83a8f/components/txn_types/src/types.rs#L118
 	actualKey = binary.BigEndian.AppendUint64(actualKey, ^w.ts)
 
 	actualValue := make([]byte, 0, 1+binary.MaxVarintLen64+1+1+len(value))
+	// below logic can be found at `WriteRef.to_bytes` [6].
+	//
+	// [6] https://github.com/tikv/tikv/blob/7793f1d5dc40206fe406ca001be1e0d7f1b83a8f/components/txn_types/src/write.rs#L362
 	actualValue = append(actualValue, 'P')
 	actualValue = binary.AppendUvarint(actualValue, w.ts)
 	actualValue = append(actualValue, 'v')
