+++ pkg/lightning/tikv/local_sst_writer_test.go
@@ -25,7 +25,13 @@ import (
 	"github.com/cockroachdb/pebble/vfs"
 	"github.com/google/uuid"
 	rockssst "github.com/lance6716/pebble/sstable"
+	"github.com/pingcap/errors"
+	"github.com/pingcap/kvproto/pkg/import_sstpb"
+	"github.com/pingcap/kvproto/pkg/kvrpcpb"
 	"github.com/stretchr/testify/require"
+	pd "github.com/tikv/pd/client"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
 )
 
 var (
@@ -34,6 +40,108 @@ var (
 	}
 	ts uint64 = 1
 )
+
+// write2ImportService4Test writes these sorted key-value pairs to the TiKV
+// cluster. SST files are generated by TiKV and saved in import directory if no
+// error happens.
+func write2ImportService4Test(
+	ctx context.Context,
+	pdAddrs []string,
+	sortedKVs [][2][]byte,
+	ts uint64,
+) ([]*import_sstpb.SSTMeta, error) {
+	pdClient, err := pd.NewClient(pdAddrs, pd.SecurityOption{})
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	defer pdClient.Close()
+
+	r0, err := pdClient.GetRegion(ctx, sortedKVs[0][0])
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	r1, err := pdClient.GetRegion(ctx, sortedKVs[len(sortedKVs)-1][0])
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	if r0.Meta.Id != r1.Meta.Id {
+		return nil, errors.Errorf(
+			"only support write to the same region, "+
+				"first key: %X, last key: %X, "+
+				"first region id: %d, last region id: %d",
+			sortedKVs[0][0], sortedKVs[len(sortedKVs)-1][0],
+			r0.Meta.Id, r1.Meta.Id,
+		)
+	}
+
+	store, err := pdClient.GetStore(ctx, r0.Leader.GetStoreId())
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	conn, err := grpc.DialContext(
+		ctx, store.GetAddress(),
+		grpc.WithTransportCredentials(insecure.NewCredentials()),
+		grpc.WithBlock(),
+	)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	defer conn.Close()
+
+	ingestClient := import_sstpb.NewImportSSTClient(conn)
+	writeStream, err := ingestClient.Write(ctx)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	u := uuid.New()
+	writeMeta := &import_sstpb.SSTMeta{
+		Uuid:        u[:],
+		RegionId:    r0.Meta.Id,
+		RegionEpoch: r0.Meta.RegionEpoch,
+		Range: &import_sstpb.Range{
+			Start: sortedKVs[0][0],
+			End:   sortedKVs[len(sortedKVs)-1][0],
+		},
+	}
+	rpcCtx := kvrpcpb.Context{
+		RegionId:    r0.Meta.Id,
+		RegionEpoch: r0.Meta.RegionEpoch,
+		Peer:        r0.Leader,
+	}
+	err = writeStream.Send(&import_sstpb.WriteRequest{
+		Chunk:   &import_sstpb.WriteRequest_Meta{Meta: writeMeta},
+		Context: &rpcCtx,
+	})
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+
+	batch := &import_sstpb.WriteBatch{
+		CommitTs: ts,
+		Pairs:    make([]*import_sstpb.Pair, 0, len(sortedKVs)),
+	}
+	for _, kv := range sortedKVs {
+		batch.Pairs = append(batch.Pairs, &import_sstpb.Pair{
+			Key:   kv[0],
+			Value: kv[1],
+		})
+	}
+	err = writeStream.Send(&import_sstpb.WriteRequest{
+		Chunk:   &import_sstpb.WriteRequest_Batch{Batch: batch},
+		Context: &rpcCtx,
+	})
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	resp, err := writeStream.CloseAndRecv()
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	if resp.GetError() != nil {
+		return nil, errors.Errorf("write failed: %s", resp.GetError())
+	}
+	return resp.Metas, nil
+}
 
 func TestGRPCWriteToTiKV(t *testing.T) {
 	t.Skip(`This is a manual test. You can use tiup playground and run this test.
